backend/requirements.txt
Flask
google-cloud-bigquery
google-cloud-core
google-auth
elasticsearch
requests
python-dotenv
flask-cors
pandas
